changed cast from long to int for printing
git-svn-id: http://internalbrr.cr.usgs.gov/svn_MOWS/PRMS_SVN/mmf/trunk@6976 6316d359-22eb-4c0b-9c0f-3bdc8ec1e293

+++ batch_run_functions.c
@@ -359,8 +359,11 @@ char *single_run_post_run () {
                            break;
 
                         case M_LONG:
+// markstro test
+//                           fprintf (ani_var_files[i], "\t%10ld",
+//                                    *((long *) var->value + j));
                            fprintf (ani_var_files[i], "\t%10ld",
-                                    *((long *) var->value + j));
+                                    *((int *) var->value + j));
                            break;
                      }
                   }
